1. moderate parameter for masks, tables of both two scenes. 2. collect part of the data from m3pi robot and apply to ctrlC.cs 3. the rest is check ctrlC's control part. (the car is down right now.)

+++ Holojam/Assets/Holojam/PhysicalSharing/Scripts/robotCtrl/SerialCommunication.cs
@@ -144,9 +144,9 @@ public class SerialCommunication {
 	public virtual void stop(){
 	}
 
-	public virtual void setSpeed(float f){}
+	public virtual void setSpeed(int f){}
 
-	public virtual  void setWaitTime(float f){}
+	public virtual  void setWaitTime(int f){}
 
 	public void right(float angle){
 		if (angle < 0)
